增加功能

+++ src/main/java/com/haiyan/deflower/service/impl/AddressServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @author haiyan
@@ -32,31 +33,46 @@ public class AddressServiceImpl implements AddressService {
     @Override
     public Long addAddress(Address address) {
         User user = userUtils.getUser(ServletUtils.getRequest());
+        if (Objects.isNull(user)) {
+            throw new ExceptionResult("Address","false",null,"请先登陆");
+        }
         address.setUid(user.getId());
         if(!addressDao.save(address)) {
-            throw new ExceptionResult("cart","false",null,"添加失败");
+            throw new ExceptionResult("Address","false",null,"添加失败");
         }
         return address.getId();
     }
 
     @Override
     public Boolean updateAddress(Address address) {
+        User user = userUtils.getUser(ServletUtils.getRequest());
+        if (Objects.isNull(user)) {
+            throw new ExceptionResult("Address","false",null,"请先登陆");
+        }
         if (!addressDao.updateById(address)) {
-            throw new ExceptionResult("cart","false",null,"修改失败");
+            throw new ExceptionResult("Address","false",null,"修改失败");
         }
         return true;
     }
 
     @Override
     public Boolean deletedAddress(Long id) {
+        User user = userUtils.getUser(ServletUtils.getRequest());
+        if (Objects.isNull(user)) {
+            throw new ExceptionResult("Address","false",null,"请先登陆");
+        }
         if (addressMapper.deleteById(id)==0) {
-            throw new ExceptionResult("cart","false",null,"删除失败");
+            throw new ExceptionResult("Address","false",null,"删除失败");
         }
         return true;
     }
 
     @Override
     public List<Address> listAddress() {
+        User user = userUtils.getUser(ServletUtils.getRequest());
+        if (Objects.isNull(user)) {
+            throw new ExceptionResult("Address","false",null,"请先登陆");
+        }
         return addressDao.lambdaQuery().eq(Address::getUid, userUtils.getUser(ServletUtils.getRequest()).getId()).list();
     }
 }
